tests for federation,operation
