+++ stellar_base/operation.py
@@ -5,6 +5,7 @@ from decimal import *
 from .asset import Asset
 from .stellarxdr import Xdr
 from .utils import account_xdr_object, signer_key_xdr_object, encode_check, best_rational_approximation as best_r, division
+from .utils import XdrLengthError
 
 ONE = Decimal(10 ** 7)
 
@@ -47,6 +48,34 @@ class Operation(object):
     def from_xdr_amount(value):
         return str(Decimal(value) / ONE)
 
+    @classmethod
+    def from_xdr(cls,xdr):
+        xdr_decode = base64.b64decode(xdr)
+        op = Xdr.StellarXDRUnpacker(xdr_decode)
+        op = op.unpack_Operation()
+        if op.type == Xdr.const.CREATE_ACCOUNT:
+            return CreateAccount.from_xdr_object(op)
+        elif op.type == Xdr.const.PAYMENT:
+            return Payment.from_xdr_object(op)
+        elif op.type == Xdr.const.PATH_PAYMENT:
+            return PathPayment.from_xdr_object(op)
+        elif op.type == Xdr.const.CHANGE_TRUST:
+            return ChangeTrust.from_xdr_object(op)
+        elif op.type == Xdr.const.ALLOW_TRUST:
+            return AllowTrust.from_xdr_object(op)
+        elif op.type == Xdr.const.SET_OPTIONS:
+            return SetOptions.from_xdr_object(op)
+        elif op.type == Xdr.const.MANAGE_OFFER:
+            return ManageOffer.from_xdr_object(op)
+        elif op.type == Xdr.const.CREATE_PASSIVE_OFFER:
+            return CreatePassiveOffer.from_xdr_object(op)
+        elif op.type == Xdr.const.ACCOUNT_MERGE:
+            return AccountMerge.from_xdr_object(op)
+        elif op.type == Xdr.const.INFLATION:
+            return Inflation.from_xdr_object(op)
+        elif op.type == Xdr.const.MANAGE_DATA:
+            return ManageData.from_xdr_object(op)
+
 
 class CreateAccount(Operation):
     def __init__(self, opts):
@@ -67,16 +96,16 @@ class CreateAccount(Operation):
         if not op_xdr_object.sourceAccount:
             source = None
         else:
-            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519)
+            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519).decode()
 
-        destination = encode_check('account', op_xdr_object.body.createAccountOp.destination.ed25519)
+        destination = encode_check('account', op_xdr_object.body.createAccountOp.destination.ed25519).decode()
         starting_balance = Operation.from_xdr_amount(op_xdr_object.body.createAccountOp.startingBalance)
 
         return cls({
             'source': source,
             'destination': destination,
             'starting_balance': starting_balance,
-        })
+            })
 
 
 class Payment(Operation):
@@ -102,9 +131,9 @@ class Payment(Operation):
         if not op_xdr_object.sourceAccount:
             source = None
         else:
-            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519)
+            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519).decode()
 
-        destination = encode_check('account', op_xdr_object.body.paymentOp.destination.ed25519)
+        destination = encode_check('account', op_xdr_object.body.paymentOp.destination.ed25519).decode()
         asset = Asset.from_xdr_object(op_xdr_object.body.paymentOp.asset)
         amount = Operation.from_xdr_amount(op_xdr_object.body.paymentOp.amount)
 
@@ -113,7 +142,7 @@ class Payment(Operation):
             'destination': destination,
             'asset': asset,
             'amount': amount,
-        })
+            })
 
 
 class PathPayment(Operation):
@@ -132,7 +161,7 @@ class PathPayment(Operation):
         dest_asset = self.dest_asset.to_xdr_object()
 
         path_payment = Xdr.types.PathPaymentOp(send_asset, Operation.to_xdr_amount(self.send_max), destination,
-                                               dest_asset, Operation.to_xdr_amount(self.dest_amount), self.path)
+                dest_asset, Operation.to_xdr_amount(self.dest_amount), self.path)
         self.body.type = Xdr.const.PATH_PAYMENT
         self.body.pathPaymentOp = path_payment
         return super(PathPayment, self).to_xdr_object()
@@ -142,9 +171,9 @@ class PathPayment(Operation):
         if not op_xdr_object.sourceAccount:
             source = None
         else:
-            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519)
+            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519).decode()
 
-        destination = encode_check('account', op_xdr_object.body.pathPaymentOp.destination.ed25519)
+        destination = encode_check('account', op_xdr_object.body.pathPaymentOp.destination.ed25519).decode()
         send_asset = Asset.from_xdr_object(op_xdr_object.body.pathPaymentOp.sendAsset)
         dest_asset = Asset.from_xdr_object(op_xdr_object.body.pathPaymentOp.destAsset)
         send_max = Operation.from_xdr_amount(op_xdr_object.body.pathPaymentOp.sendMax)
@@ -163,7 +192,7 @@ class PathPayment(Operation):
             'dest_asset': dest_asset,
             'dest_amount': dest_amount,
             'path': path
-        })
+            })
 
 
 class ChangeTrust(Operation):
@@ -189,7 +218,7 @@ class ChangeTrust(Operation):
         if not op_xdr_object.sourceAccount:
             source = None
         else:
-            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519)
+            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519).decode()
 
         line = Asset.from_xdr_object(op_xdr_object.body.changeTrustOp.line)
         print(line)
@@ -199,7 +228,7 @@ class ChangeTrust(Operation):
             'source': source,
             'asset': line,
             'limit': limit
-        })
+            })
 
 
 class AllowTrust(Operation):
@@ -235,8 +264,8 @@ class AllowTrust(Operation):
         if not op_xdr_object.sourceAccount:
             source = None
         else:
-            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519)
-        trustor = encode_check('account', op_xdr_object.body.allowTrustOp.trustor.ed25519)
+            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519).decode()
+        trustor = encode_check('account', op_xdr_object.body.allowTrustOp.trustor.ed25519).decode()
         authorize = op_xdr_object.body.allowTrustOp.authorize
 
         asset_type = op_xdr_object.body.allowTrustOp.asset.type
@@ -252,7 +281,7 @@ class AllowTrust(Operation):
             'trustor': trustor,
             'authorize': authorize,
             'asset_code': asset_code
-        })
+            })
 
 
 class SetOptions(Operation):
@@ -266,7 +295,7 @@ class SetOptions(Operation):
         self.med_threshold = opts.get('med_threshold')
         self.high_threshold = opts.get('high_threshold')
         self.home_domain = opts.get('home_domain')
-        
+
         self.signer_address = opts.get('signer_address')
         self.signer_type = opts.get('signer_type')
         self.signer_weight = opts.get('signer_weight')
@@ -277,11 +306,11 @@ class SetOptions(Operation):
             except DecodeError:
                 raise Exception('must be a valid strkey if not give signer_type')
             self.signer_type = 'ed25519PublicKey'
-            
+
         if self.signer_type in ('hashX','preAuthTx') and \
-            (self.signer_address is None or len(self.signer_address) != 32):
-            raise Exception('hashX or preAuthTx Signer must be 32 bytes')
-        
+                (self.signer_address is None or len(self.signer_address) != 32):
+                    raise Exception('hashX or preAuthTx Signer must be 32 bytes')
+
         if self.signer_type is not None and self.signer_type not in ('ed25519PublicKey','hashX','preAuthTx'):
             raise Exception('invalid signer type.')
 
@@ -309,15 +338,15 @@ class SetOptions(Operation):
         self.home_domain = assert_option_array(self.home_domain)
 
         if self.signer_address is not None and \
-            self.signer_type is not None and \
-            self.signer_weight is not None:
-            signer = [Xdr.types.Signer(signer_key_xdr_object(self.signer_type,self.signer_address), self.signer_weight)]
+                self.signer_type is not None and \
+                self.signer_weight is not None:
+                    signer = [Xdr.types.Signer(signer_key_xdr_object(self.signer_type,self.signer_address), self.signer_weight)]
         else:
             signer = []
 
         set_options_op = Xdr.types.SetOptionsOp(inflation_dest, self.clear_flags, self.set_flags,
-                                                self.master_weight, self.low_threshold, self.med_threshold,
-                                                self.high_threshold, self.home_domain, signer)
+                self.master_weight, self.low_threshold, self.med_threshold,
+                self.high_threshold, self.home_domain, signer)
         self.body.type = Xdr.const.SET_OPTIONS
         self.body.setOptionsOp = set_options_op
         return super(SetOptions, self).to_xdr_object()
@@ -327,12 +356,12 @@ class SetOptions(Operation):
         if not op_xdr_object.sourceAccount:
             source = None
         else:
-            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519)
+            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519).decode()
 
         if not op_xdr_object.body.setOptionsOp.inflationDest:
             inflation_dest = None
         else:
-            inflation_dest = encode_check('account', op_xdr_object.body.setOptionsOp.inflationDest[0].ed25519)
+            inflation_dest = encode_check('account', op_xdr_object.body.setOptionsOp.inflationDest[0].ed25519).decode()
 
         clear_flags = op_xdr_object.body.setOptionsOp.clearFlags  # list
         set_flags = op_xdr_object.body.setOptionsOp.setFlags
@@ -345,7 +374,7 @@ class SetOptions(Operation):
         if op_xdr_object.body.setOptionsOp.signer:
             key = op_xdr_object.body.setOptionOp.signer[0].key
             if key.type == Xdr.const.SIGNER_KEY_TYPE_ED25519:
-                signer_address = encode_check('account', key.ed25519)
+                signer_address = encode_check('account', key.ed25519).decode()
                 signer_type = 'ed25519PublicKey'
             if key.type == Xdr.const.SIGNER_KEY_TYPE_PRE_AUTH_TX:
                 signer_address = key.preAuthTx
@@ -373,7 +402,7 @@ class SetOptions(Operation):
             'signer_address': signer_address,
             'Signer_type': signer_type,
             'signer_weight': signer_weight
-        })
+            })
 
 
 class ManageOffer(Operation):
@@ -403,7 +432,7 @@ class ManageOffer(Operation):
         if not op_xdr_object.sourceAccount:
             source = None
         else:
-            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519)
+            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519).decode()
 
         selling = Asset.from_xdr_object(op_xdr_object.body.manageOfferOp.selling)
         buying = Asset.from_xdr_object(op_xdr_object.body.manageOfferOp.buying)
@@ -421,7 +450,7 @@ class ManageOffer(Operation):
             'amount': amount,
             'price': price,
             'offer_id': offer_id
-        })
+            })
 
 
 class CreatePassiveOffer(Operation):
@@ -451,7 +480,7 @@ class CreatePassiveOffer(Operation):
         if not op_xdr_object.sourceAccount:
             source = None
         else:
-            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519)
+            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519).decode()
 
         selling = Asset.from_xdr_object(op_xdr_object.body.createPassiveOfferOp.selling)
         buying = Asset.from_xdr_object(op_xdr_object.body.createPassiveOfferOp.buying)
@@ -466,7 +495,7 @@ class CreatePassiveOffer(Operation):
             'buying': buying,
             'amount': amount,
             'price': price
-        })
+            })
 
 
 class AccountMerge(Operation):
@@ -486,14 +515,14 @@ class AccountMerge(Operation):
         if not op_xdr_object.sourceAccount:
             source = None
         else:
-            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519)
+            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519).decode()
 
-        destination = encode_check('account', op_xdr_object.body.destination.ed25519)
+        destination = encode_check('account', op_xdr_object.body.destination.ed25519).decode()
 
         return cls({
             'source': source,
             'destination': destination
-        })
+            })
 
 
 class Inflation(Operation):
@@ -509,7 +538,7 @@ class Inflation(Operation):
         if not op_xdr_object.sourceAccount:
             source = None
         else:
-            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519)
+            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519).decode()
         return cls({'source': source})
 
 
@@ -518,6 +547,8 @@ class ManageData(Operation):
         super(ManageData, self).__init__(opts)
         self.data_name = opts.get('data_name')
         self.data_value = opts.get('data_value')
+        if len(self.data_name) > 64 or (self.data_value is not None and len(self.data_value) > 64):
+            raise XdrLengthError("Data or value should be <= 64 bytes (ascii encoded). ")
 
     def to_xdr_object(self):
         data_name = bytearray(self.data_name, encoding='utf-8')
@@ -536,7 +567,7 @@ class ManageData(Operation):
         if not op_xdr_object.sourceAccount:
             source = None
         else:
-            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519)
+            source = encode_check('account', op_xdr_object.sourceAccount[0].ed25519).decode()
 
         data_name = op_xdr_object.body.manageDataOp.dataName.decode()
 
@@ -548,4 +579,4 @@ class ManageData(Operation):
             'source': source,
             'data_name': data_name,
             'data_value': data_value
-        })
+            })
